// src/interfaces/category-interface.ts
// Generated by https://quicktype.io

export interface Categories {
    data: Category[];
    meta: Meta;
}

export interface Category {
    id:         number;
    attributes: CategoryAtributes;
}

export interface CategoryAtributes {
    name:        string;
    createdAt:   string;
    updatedAt:   string;
    publishedAt: string;
    img:         Img;
    products:    Products;
}

export interface Img {
    data: Data | null;
}

export interface Data {
    id:         number;
    attributes: DataAttributes;
}

export interface DataAttributes {
    name:              string;
    alternativeText:   null;
    caption:           null;
    width:             number;
    height:            number;
    formats:           Formats;
    hash:              string;
    ext:               string;
    mime:              string;
    size:              number;
    url:               string;
    previewUrl:        null;
    provider:          string;
    provider_metadata: ProviderMetadata;
    createdAt:         string;
    updatedAt:         string;
}

export interface Formats {
    thumbnail: Large;
    small:     Large;
    large:     Large;
    medium:    Large;
}

export interface Large {
    name:              string;
    hash:              string;
    ext:               string;
    mime:              string;
    path:              null;
    width:             number;
    height:            number;
    size:              number;
    url:               string;
    provider_metadata: ProviderMetadata;
}

export interface ProviderMetadata {
    public_id:     string;
    resource_type: ResourceType;
}

export enum ResourceType {
    Image = "image",
}

export interface Products {
    data: Product[];
}

export interface Product {
    id:         number;
    attributes: ProductAtributes;
}

export interface ProductAtributes {
    name:        string;
    Available:   boolean;
    stock:       number;
    price:       number;
    createdAt:   string;
    updatedAt:   string;
    publishedAt: string;
    description: string;
}

export interface Meta {
    pagination: Pagination;
}

export interface Pagination {
    page:      number;
    pageSize:  number;
    pageCount: number;
    total:     number;
}